+++ src/main/java/org/bytedeco/javacpp/BytePointer.java
@@ -78,8 +78,8 @@ public class BytePointer extends Pointer {
         super(buffer);
         if (buffer != null && buffer.hasArray()) {
             byte[] array = buffer.array();
-            allocateArray(array.length);
-            put(array);
+            allocateArray(array.length - buffer.arrayOffset());
+            put(array, buffer.arrayOffset(), array.length - buffer.arrayOffset());
             position(buffer.position());
             limit(buffer.limit());
         }