+++ src/main/java/org/bytedeco/javacpp/DoublePointer.java
@@ -52,8 +52,8 @@ public class DoublePointer extends Pointer {
         super(buffer);
         if (buffer != null && buffer.hasArray()) {
             double[] array = buffer.array();
-            allocateArray(array.length);
-            put(array);
+            allocateArray(array.length - buffer.arrayOffset());
+            put(array, buffer.arrayOffset(), array.length - buffer.arrayOffset());
             position(buffer.position());
             limit(buffer.limit());
         }