+++ src/main/java/org/bytedeco/javacpp/IntPointer.java
@@ -62,8 +62,8 @@ public class IntPointer extends Pointer {
         super(buffer);
         if (buffer != null && buffer.hasArray()) {
             int[] array = buffer.array();
-            allocateArray(array.length);
-            put(array);
+            allocateArray(array.length - buffer.arrayOffset());
+            put(array, buffer.arrayOffset(), array.length - buffer.arrayOffset());
             position(buffer.position());
             limit(buffer.limit());
         }